% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{build_database}
\alias{build_database}
\title{Build and optionally save a database}
\usage{
build_database(path, pattern = "^.*\\\\.tif$", save_db = FALSE)
}
\arguments{
\item{path}{char, path to the dataset}

\item{pattern}{char pattern to search for (as regular expression)}

\item{save_db}{logical, if TRUE save the database in the specified path as database.csv.gz}
}
\description{
Build and optionally save a database
}
