% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{current_month}
\alias{current_month}
\title{Retrieve the current month}
\usage{
current_month(x = Sys.Date(), offset = 0)
}
\arguments{
\item{x}{Date the date to compute upon (useful for "offset")}

\item{offset}{numeric, days to offset. If 0 (the default) then the current month, 
if -1 then the prior month, to 2 months before try -32, to get the next month try
32.  It's a hack.}
}
\value{
Date class object
}
\description{
Retrieve the current month
}
