% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{extract_points}
\alias{extract_points}
\title{Extract point data from a stars object}
\usage{
extract_points(x, y, pnames = NULL, form = c("long", "wide")[1])
}
\arguments{
\item{x}{stars object (with or without time dimension)}

\item{y}{sf point data}

\item{pnames}{NULL or chr, names to apply to the output point data. If NULL then
names are assinged as "p01" or "p001" etc depending upon the number of input points.}

\item{form}{chr one of either "long" (the default) or "wide" to control 
output column layout.}
}
\value{
table of variable data for each input point
}
\description{
Extract point data from a stars object
}
