% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_day}
\alias{fetch_day}
\title{Fetch and store one or more dates}
\usage{
fetch_day(
  dates = Sys.Date() + c(-10, -9, -8),
  param = "sst.day.mean",
  bb = c(xmin = 0, ymin = -90, xmax = 360, ymax = 90),
  path = oisst_path("world")
)
}
\arguments{
\item{dates}{Date, the dates}

\item{param}{char, the parameter to fetch}

\item{bb}{numeric, named vector of xmin, xmax, ymin and ymax bounding box}

\item{path}{char, the output path}
}
\value{
tibble database (possibly with no wrows)
}
\description{
Fetch and store one or more dates
}
