% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_ltm}
\alias{fetch_ltm}
\title{Fetch long time series means}
\usage{
fetch_ltm(
  param = "sst.mon.ltm.1991-2020",
  bb = c(xmin = 0, ymin = 0, xmax = 360, ymax = 90),
  path = oisst_path("world")
)
}
\arguments{
\item{param}{char the parameter}

\item{bb}{numeric 4 element bounding box in the range of xmin = 0, ymin = -90, xmax= 360, ymax = 90}

\item{path}{char, the output path}
}
\value{
tibble database for existing files
}
\description{
Fetch long time series means
}
