% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_mask}
\alias{fetch_mask}
\title{Fetch the land-sea mask}
\usage{
fetch_mask(bb = c(xmin = 0, ymin = 0, xmax = 360, ymax = 90))
}
\arguments{
\item{bb}{numeric bounding box for subsetting}
}
\value{
stars object
}
\description{
Fetch the land-sea mask
}
