% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_oisst}
\alias{fetch_oisst}
\title{Fetch OISST data for a given time range, parameter and bounding box.}
\usage{
fetch_oisst(
  dates = c(as.Date("1991-09-01"), Sys.Date()),
  param = "sst.day.mean",
  bb = c(xmin = 0, ymin = -90, xmax = 360, ymax = 90),
  path = oisst_path("world")
)
}
\arguments{
\item{dates}{Date, two element vector of dates specifying the start and stop range}

\item{param}{char, identifies the parameter including
\itemize{
 \item{sst.day.mean}
 \item{icec.day.mean}
 \item{sst.week.mean}
 \item{icec.week.mean}
 \item{sst.month.mean}
 \item{icec.month.mean}
 \item{sst.day.mean.ltm.<start>-<stop>  include stars and stop years}
 \item{icec.day.mean.ltm.<start>-<stop> include stars and stop years}
}}

\item{bb}{numeric, named vector of xmin, xmax, ymin and ymax}

\item{path}{char the root path where data is saved}
}
\description{
Fetch OISST data for a given time range, parameter and bounding box.
}
