% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_week}
\alias{fetch_week}
\title{Fetch one or more weeks}
\usage{
fetch_week(
  dates = seq(from = as.Date("1981-09-06"), to = Sys.Date() - 1, by = "7 days"),
  param = "sst.week.mean",
  bb = c(xmin = 0, ymin = 0, xmax = 360, ymax = 90),
  path = oisst_path("world")
)
}
\arguments{
\item{dates}{two element range of dates}

\item{param}{char, the identifier of the data set}

\item{bb}{numeric 4 element bounding box in the range of 
(xmin = 0, ymin = -90, xmax= 360, ymax = 90)}

\item{path}{char, the output path}
}
\value{
tibble database for existing files
}
\description{
Fetch one or more weeks
}
