% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_year}
\alias{fetch_year}
\title{Fetch and store an entire year}
\usage{
fetch_year(year = 1981, param = "sst.day.mean", path = oisst_path("world"))
}
\arguments{
\item{year}{num, the year to fetch}

\item{param}{char, the parameter to fetch}

\item{path}{char, the output path}
}
\value{
tibble database of existing files
}
\description{
Fetch and store an entire year
}
