% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{generate_filename}
\alias{generate_filename}
\title{Generate filenames for ncdf4 contents}
\usage{
generate_filename(x, dates = get_time(x), format = "\%Y-\%m-\%d", ext = ".tif")
}
\arguments{
\item{x}{ncdf4 object}

\item{dates}{Date, the dates to format, by default all available}

\item{format}{char, the date format to use}

\item{ext}{char extension to apply to filename (default ".tif")}
}
\value{
char vector of "param.per.trt.YYYY-mm-dd.ext"
  or "param.per.trt.YYYY-YYYY.YYYY-mm-dd.ext" for ltm
}
\description{
Generate filenames for ncdf4 contents
}
