% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{get_epoch}
\alias{get_epoch}
\title{Retrieve the initial epoch as Date class}
\usage{
get_epoch(x)
}
\arguments{
\item{x}{ncdf4 object}
}
\value{
Date class object
}
\description{
Retrieve the initial epoch as Date class
}
