% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{get_geometry}
\alias{get_geometry}
\title{Retrieve the a list of geometry info}
\usage{
get_geometry(x)
}
\arguments{
\item{x}{ncdf4 object}
}
\value{
list of [x/y] limits and [x/y] steps
}
\description{
Retrieve the a list of geometry info
}
