% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{get_nav}
\alias{get_nav}
\title{Get the navigation list for extracting form the ncdf object}
\usage{
get_nav(
  x,
  varid = get_varname(x),
  bb = c(xmin = 0, ymin = -90, xmax = 360, ymax = 90),
  time = c(1, 1)
)
}
\arguments{
\item{x}{ncfd object}

\item{varid}{char the name of the variable to be extracted}

\item{bb}{numeric, named vector of xmin, xmax, ymin and ymax for extraction}

\item{time}{numeric, 2 element time indices (or Dates over which to extract)
in the form (start, count) or (startDate, endDate)}
}
\value{
a navigation structure with
\itemize{
  \item{varid the id of the variable}
  \item{bb the requested bounding box}
  \item{bbox the actual bounding box}
  \item{ndim the number of dimensions}
  \item{start numeric vector of start indices}
  \item{count numeric vector of extraction sizes}
  }
}
\description{
Get the navigation list for extracting form the ncdf object
}
