% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{get_one}
\alias{get_one}
\title{Retrieve a single layer (date) for one attribute}
\usage{
get_one(x, att = names(x$var)[1], time = get_time(x)[1], left = c(0, -180)[1])
}
\arguments{
\item{x}{ncdf4 object}

\item{att}{char, the name of the attribute (variable)}

\item{time}{Date or index of the date to retrieve}

\item{left}{numeric one of 0 (default) or -180.  If the latter the x values
span [-180, 180] otherwise [0,360]}
}
\value{
stars object
}
\description{
Retrieve a single layer (date) for one attribute
}
