% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{get_time}
\alias{get_time}
\title{Retrieve the time values as Date class}
\usage{
get_time(x, epoch = get_epoch(x))
}
\arguments{
\item{x}{ncdf4 object}

\item{epoch}{Date the epoch from which time is figured}
}
\value{
Date class vector
}
\description{
Retrieve the time values as Date class
}
