% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{get_time_index}
\alias{get_time_index}
\title{Get the indices for specified times}
\usage{
get_time_index(x, time)
}
\arguments{
\item{x}{ncdf4 object}

\item{time}{Date class times}
}
\value{
index of where the specified time falls relative to the times in \code{x}
}
\description{
Get the indices for specified times
}
