% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{get_varname}
\alias{get_varname}
\title{Get ncdf variable names and return the most likely candidate.}
\usage{
get_varname(x, likely = c("sst", "icec"))
}
\arguments{
\item{x}{ncdf4 object}

\item{likely}{char, one or more most likely candidates}
}
\value{
character variable name
}
\description{
If most likely candidate is not found then the first is returned
}
