% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{list_files}
\alias{list_files}
\title{List files in a database}
\usage{
list_files(path, pattern = "^.*\\\\.tif$", ...)
}
\arguments{
\item{path}{char, path to the dataset}

\item{pattern}{char pattern to search for (as regular expression)}

\item{...}{other arguments for `list.files`}
}
\value{
character vector of zero or more filename
}
\description{
List files in a database
}
