% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{list_oisst}
\alias{list_oisst}
\title{List the data sets offered by OISST thredds catalog}
\usage{
list_oisst(base_uri = oisst_uri("xml"))
}
\arguments{
\item{base_uri}{char, the base URI to the top level catalog (as .xml)}
}
\value{
dataframe of available resources
}
\description{
List the data sets offered by OISST thredds catalog
}
