% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{match_weeks}
\alias{match_weeks}
\title{Match a range of dates to the the sequence of weekly dates in OISST}
\usage{
match_weeks(dates)
}
\arguments{
\item{dates}{a two element Date vector of start and stop}
}
\value{
a sequence of dates
}
\description{
Match a range of dates to the the sequence of weekly dates in OISST
}
