% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{oisst_path}
\alias{oisst_path}
\title{Retrieve a OISST path}
\usage{
oisst_path(..., root = root_path())
}
\arguments{
\item{...}{file segements to add to the root path. See \code{\link[base]{file.path}}}

\item{root}{the root path to the OISST data}
}
\value{
character path description
}
\description{
Retrieve a OISST path
}
