% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{oisst_uri}
\alias{oisst_uri}
\title{Retrieve the base URL for the catralog as HTML or XML}
\usage{
oisst_uri(what = c("html", "xml")[1])
}
\arguments{
\item{what}{char one of 'html' or 'xml' (default)}
}
\value{
URL
}
\description{
Retrieve the base URL for the catralog as HTML or XML
}
