% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_oisst}
\alias{query_oisst}
\title{Query the OISST thredds catalog}
\usage{
query_oisst(
  year = seq(from = 1981, to = as.numeric(format(Sys.Date(), "\%Y")), by = 1),
  param = "sst.day.mean",
  base_uri = oisst_uri("xml")
)
}
\arguments{
\item{year}{num, one or more years to query}

\item{param}{char, one parameter to query}

\item{base_uri}{char, the base URI to the top level catalog (as .xml)}
}
\value{
named list of URLs to datastes matching the query
}
\description{
Query the OISST thredds catalog
}
