% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{read_database}
\alias{read_database}
\title{Read a OISST database}
\usage{
read_database(path, filename = "database.csv.gz")
}
\arguments{
\item{path}{char, path description to the database}

\item{filename}{char, name of the file in the path}
}
\description{
Read a OISST database
}
