% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_mask}
\alias{read_mask}
\title{Read an OISST mask}
\usage{
read_mask(x = "world", path = oisst_path("mask"))
}
\arguments{
\item{x}{char, the name of the mask to read (default "world")}

\item{path}{char the path to the mask data}
}
\value{
stars class object
}
\description{
Read an OISST mask
}
