% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_oisst}
\alias{read_oisst}
\title{Read OISST data files}
\usage{
read_oisst(x, path = oisst_path(), along = NULL)
}
\arguments{
\item{x}{database (tibble) of common param, per type (no checking done)}

\item{path}{char, the path to the data}

\item{along}{NULL or somehting user specified.  If NULL then 
we assume the user wants to stack by date, but the user might 
want one attribute per database row.}
}
\value{
stars object
}
\description{
Read OISST data files
}
