% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{root_path}
\alias{root_path}
\title{Get the core data path from a user specified file}
\usage{
root_path(filename = "~/.oisstdata")
}
\arguments{
\item{filename}{the name the file to store the path as a single line of text}
}
\value{
character data path
}
\description{
Get the core data path from a user specified file
}
