% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{to_180_bb}
\alias{to_180_bb}
\title{Convert a bounding box to -180-180}
\usage{
to_180_bb(x)
}
\arguments{
\item{x}{four element named vector of 'xmin', 'ymin', 'xmax' and 'ymax'}
}
\value{
a four element vector of 'xmin', 'ymin', 'xmax' and 'ymax'
}
\description{
Convert a bounding box to -180-180
}
