% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{write_database}
\alias{write_database}
\title{Write a OISST database}
\usage{
write_database(x, path, filename = "database.csv.gz")
}
\arguments{
\item{x}{tibble database}

\item{path}{char, path description to the database}

\item{filename}{cahr, the filename to save to}
}
\value{
the input database
}
\description{
Write a OISST database
}
