% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{compose_filename}
\alias{compose_filename}
\title{Given a database and path, compose filenames}
\usage{
compose_filename(x = decompose_filename(), path = "/foo/bar", ext = "tif")
}
\arguments{
\item{x}{tibble database}

\item{path}{char, path description}

\item{ext}{char the extension for the files}
}
\value{
fully qualified paths
}
\description{
Given a database and path, compose filenames
}
