% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{decompose_filename}
\alias{decompose_filename}
\title{Given one or more filename, decompose into a database}
\usage{
decompose_filename(
  x = c("sst.day.mean.1981-09-01.tif", "icec.mon.max.1981-09-01.tif",
    "icec.mon.max.ltm.1991-2020.tif", "sst.day.mean.ltm.1971-2000.1971-01-01.tif",
    "sst.mon.ltm.1991-2020.1991-11-01.tif"),
  ext = ".tif"
)
}
\arguments{
\item{x}{char vector of filenames}

\item{ext}{char the extension to remove before processing. NA to skip.}
}
\value{
tibble database
}
\description{
Given one or more filename, decompose into a database
}
