% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_month}
\alias{fetch_month}
\title{Fetch a series of months}
\usage{
fetch_month(
  date = seq(from = as.Date("1981-01-01"), to = current_month(), by = "month"),
  bb = c(xmin = 0, ymin = 0, xmax = 360, ymax = 90),
  path = oisst_path("world")
)
}
\arguments{
\item{date}{Date, one or more dates}

\item{bb}{numeric 4 element bounding box in the range of 
(xmin = 0, ymin = -90, xmax= 360, ymax = 90)}

\item{path}{char, the output path}

\item{logical, }{TRUE if successful}
}
\description{
Fetch a series of months
}
