% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{get_one}
\alias{get_one}
\title{Retrieve a single layer (date) for one attribute}
\usage{
get_one(
  x,
  att = get_varname(x),
  bb = c(xmin = 0, ymin = -90, xmax = 360, ymax = 90),
  time = 1,
  nav = NULL
)
}
\arguments{
\item{x}{ncdf4 object}

\item{att}{char, the name of the attribute (variable)}

\item{bb}{numeric, named vector of xmin, xmax, ymin and ymax}

\item{time}{Date or index of the date to retrieve}

\item{nav}{list of navigation info, or NULL}
}
\value{
stars object
}
\description{
Retrieve a single layer (date) for one attribute
}
