% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_oisst}
\alias{read_oisst}
\title{Read OISST data files}
\usage{
read_oisst(x, path)
}
\arguments{
\item{x}{database (tibble) of common param, per type (no checking done)}

\item{path}{char, the path to the data}
}
\value{
stars object
}
\description{
Read OISST data files
}
